+++ packages/orq-rc/src/models/components/evaluatorresponsejsonschema.ts
@@ -32,8 +32,8 @@ export const EvaluatorResponseJsonSchema$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-20T13:19:51.038Z"),
-  updated: z.string().default("2026-08-20T13:19:51.038Z"),
+  created: z.string().default("2026-08-20T15:20:52.764Z"),
+  updated: z.string().default("2026-08-20T15:20:52.764Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),